% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_delta_iso.R
\name{calcDeltas}
\alias{calcDeltas}
\title{Calculate Delta Values Function}
\usage{
calcDeltas(ysim)
}
\arguments{
\item{ysim}{The NANNO model simulation results.}
}
\description{
This function takes model simulations and calculates delta values.
}
\examples{
calcDeltas(ysim)
}
\keyword{NANNO}

