% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcMasses.R
\name{calcMasses}
\alias{calcMasses}
\title{Calculate Masses For Each Process}
\usage{
calcMasses(NANNOmodel, ysim)
}
\arguments{
\item{NANNOmodel}{The NANNO model object, typically after model fitting.}

\item{ysim}{The NANNO model simulation results.}
}
\description{
This function takes model results from a NANNOmodel via parms() and simulation values from ysim and calculates the mass of nitrogen moved by each process
}
\examples{
calcMasses(NANNOmodel, ysim)
}
\keyword{NANNO}

