% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ratio_delta.R
\name{deltaToRatio}
\alias{deltaToRatio}
\title{Ratio To Delta Function}
\usage{
deltaToRatio(delta)
}
\arguments{
\item{delta}{The delta15N values.}
}
\description{
This function converts delta values into isotopic ratios for nitrogen.
}
\examples{
deltaToRatio(delta)
}
\keyword{NANNO}

