% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NANNO_fit.R
\name{NANNO_fit}
\alias{NANNO_fit}
\title{NANNO Fitting Function}
\usage{
NANNO_fit(filename)
}
\arguments{
\item{filename}{The base name of the files to be used in NANNO model fitting.}
}
\description{
This function takes field data and iteratively fits
}
\examples{
NANNO_fit(filename)
}
\keyword{NANNO}

