% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NANNO_fit_params.R
\name{NANNO_fit_params}
\alias{NANNO_fit_params}
\title{NANNO Params Function}
\usage{
NANNO_fit_params(NANNOmodel, whichpar)
}
\arguments{
\item{NANNOmodel}{The NANNO model object, typically after model fitting.}

\item{whichpar}{The parameters that are actively fit.}
}
\description{
This function takes model simulations and observations and calculates fitting statistics.
}
\examples{
NANNO_fit_params(NANNOmodel, whichpar)
}
\keyword{NANNO}
