% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NANNO_fit_stats.R
\name{NANNO_fit_stats}
\alias{NANNO_fit_stats}
\title{NANNO Statistics Function}
\usage{
NANNO_fit_stats(ysim, yobs, obstime)
}
\arguments{
\item{ysim}{The NANNO model simulation results.}

\item{yobs}{The field observations.}

\item{obstime}{Field observation times}
}
\description{
This function takes model simulations and observations and calculates fitting statistics.
}
\examples{
NANNO_fit_stats(ysim, yobs, obstime)
}
\keyword{NANNO}
