% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NANNO_input_converting.R
\name{NANNO_input_converting}
\alias{NANNO_input_converting}
\title{NANNO Inputs Function}
\usage{
NANNO_input_converting(filename)
}
\arguments{
\item{filename}{The base name of the bounds file to be used.}
}
\description{
This function takes inputs for lower and upper bounds from a csv file and converts them for the fitting routine
}
\examples{
NANNO_input_converting(filename)
}
\keyword{NANNO}

