% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NANNO_test.R
\name{NANNO_test}
\alias{NANNO_test}
\title{Test the NANNO Function}
\usage{
NANNO_test()
}
\description{
This function takes model results from the default NANNO_model() and creates a default figure
}
\examples{
NANNO_test()
}
\keyword{NANNO}
