% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ratio_delta.R
\name{ratioToDelta}
\alias{ratioToDelta}
\title{Ratio To Delta Function}
\usage{
ratioToDelta(ratio)
}
\arguments{
\item{ratio}{The 15N/14N isotopic ratios.}
}
\description{
This function converts isotopic ratios into delta values for nitrogen.
}
\examples{
ratioToDelta(ratio)
}
\keyword{NANNO}

