% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residuals_figures.R
\name{residuals_figures}
\alias{residuals_figures}
\title{Two Part Figure with Observations Function}
\usage{
residuals_figures(ysim, yobs, obstime)
}
\arguments{
\item{ysim}{The NANNO model simulation results.}

\item{yobs}{The field observations.}

\item{obstime}{Field observation times.}
}
\description{
This function takes model simulations and observations, calculate the residuals and plots them.
}
\examples{
residuals_figures(ysim, yobs, obstime)
}
\keyword{NANNO}
