% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TAN_speciation.R
\name{TAN_speciation}
\alias{TAN_speciation}
\title{TAN Speciation Function}
\usage{
TAN_speciation(TAN, pH)
}
\arguments{
\item{TAN}{The TAN concentration.}

\item{pH}{The pH.}
}
\description{
This function takes TAN concentration and pH and calculates/returns the concentration of NH3 and NH4+ as a vector.
}
\examples{
TAN_speciation(TAN, pH)
}
\keyword{NANNO}
