% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two_part_figures.R
\name{two_part_figure}
\alias{two_part_figure}
\title{Two Part Figure Function}
\usage{
two_part_figure(ysim)
}
\arguments{
\item{ysim}{The NANNO model simulation results.}
}
\description{
This function takes model simulations and plots them with 3 concentrations (NO3-, TAN, DIN) on one row and isotopes on another.
}
\examples{
two_part_figure(ysim)
}
\keyword{NANNO}

