% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two_part_figures.R
\name{two_part_figure_all}
\alias{two_part_figure_all}
\title{Two Part Figure Function with All Stocks}
\usage{
two_part_figure_all(ysim)
}
\arguments{
\item{ysim}{The NANNO model simulation results.}
}
\description{
This function takes model simulations and plots them with all 4 concentrations (NO3-, TAN, N2O, DIN) on one row and isotopes on another.
Colours are from scales::hue_pal()(5) to use the same colour for all 5 potential variables
"#F8766D" "#A3A500" "#00BF7D" "#00B0F6" "#E76BF3"
See https://github.com/hadley/scales/blob/master/R/pal-hue.r
}
\examples{
two_part_figure(ysim)
}
\keyword{NANNO}
