% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two_part_figures.R
\name{two_part_figure_with_obs}
\alias{two_part_figure_with_obs}
\title{Two Part Figure with Observations Function}
\usage{
two_part_figure_with_obs(ysim, yobs, obstime)
}
\arguments{
\item{ysim}{The NANNO model simulation results.}

\item{yobs}{The field observations.}

\item{obstime}{Field observation times}
}
\description{
This function takes model simulations and observations and plots them with concentrations on one row and isotopes on another.
}
\examples{
two_part_figure_with_obs(ysim, yobs)
}
\keyword{NANNO}

