% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two_part_figures.R
\name{two_part_figure_all}
\alias{two_part_figure_all}
\title{Two Part Figure Function with All Stocks}
\usage{
two_part_figure_all(ysim)
}
\arguments{
\item{ysim}{The NANNO model simulation results.}
}
\description{
This function takes model simulations and plots them with all 5 concentrations (NO2- NO3-, TAN, N2O, DIN) on one row and isotopes on another.
}
\examples{
two_part_figure(ysim)
}
\keyword{NANNO}

