% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two_part_figures.R
\name{two_part_figure_with_obs}
\alias{two_part_figure_with_obs}
\title{Two Part Figure with Observations Function}
\usage{
two_part_figure_with_obs(ysim, yobs, obstime)
}
\arguments{
\item{ysim}{The NANNO model simulation results.}

\item{yobs}{The field observations.}

\item{obstime}{Field observation times}
}
\description{
This function takes model simulations and observations and plots them with concentrations on one row and isotopes on another.
Colours are from scales::hue_pal()(5) to use the same colour for all 5 potential variables
"#F8766D" "#A3A500" "#00BF7D" "#00B0F6" "#E76BF3"
See https://github.com/hadley/scales/blob/master/R/pal-hue.r
}
\examples{
two_part_figure_with_obs(ysim, yobs)
}
\keyword{NANNO}
